Add JWT authentication and enforce user-specific authorization for chat threads

+++ Backend/routes/chat.js
@@ -1,6 +1,7 @@
 import express from "express";
 import Thread from "../models/Thread.js";
 import getOpenAIAPIResponse from "../utils/groqai.js";
+import authMiddleware from "../middleware/authMiddleware.js";
 
 const router = express.Router();
 
@@ -21,9 +22,9 @@ router.post("/test", async (req, res) => {
 });
 
 //Get all threads
-router.get("/thread", async (req, res) => {
+router.get("/thread", authMiddleware, async (req, res) => {
   try {
-    const threads = await Thread.find({}).sort({ updatedAt: -1 });
+    const threads = await Thread.find({userId: req.user.userId}).sort({ updatedAt: -1 });
     //descending order of updatedAt...most recent data on top
     res.json(threads);
   } catch (err) {
@@ -32,11 +33,11 @@ router.get("/thread", async (req, res) => {
   }
 });
 
-router.get("/thread/:threadId", async (req, res) => {
+router.get("/thread/:threadId", authMiddleware, async (req, res) => {
   const { threadId } = req.params;
 
   try {
-    const thread = await Thread.findOne({ threadId });
+    const thread = await Thread.findOne({ threadId,userId: req.user.userId });
 
     if (!thread) {
       res.status(404).json({ error: "Thread not found" });
@@ -49,11 +50,11 @@ router.get("/thread/:threadId", async (req, res) => {
   }
 });
 
-router.delete("/thread/:threadId", async (req, res) => {
+router.delete("/thread/:threadId",  authMiddleware,async (req, res) => {
   const { threadId } = req.params;
 
   try {
-    const deletedThread = await Thread.findOneAndDelete({ threadId });
+    const deletedThread = await Thread.findOneAndDelete({ threadId ,userId: req.user.userId});
 
     if (!deletedThread) {
       res.status(404).json({ error: "Thread not found" });
@@ -66,10 +67,11 @@ router.delete("/thread/:threadId", async (req, res) => {
   }
 });
 
-router.post("/chat", async (req, res) => {
+router.post("/chat", authMiddleware, async (req, res) => {
   // console.log("BODY:", req.body);
   const { threadId, message } = req.body;
-
+  const userId = req.user.userId;
+  
   if (!threadId || !message) {
     res.status(400).json({ error: "missing required fields" });
   }
@@ -82,6 +84,7 @@ router.post("/chat", async (req, res) => {
       thread = new Thread({
         threadId,
         title: message,
+        userId,
         messages: [{ role: "user", content: message }],
       });
     } else {
